# profileService/pkg/api/openapi.yaml
openapi: 3.0.3
info:
  title: Profile Service API
  description: |
    API сервиса управления профилями преподавателей и курсами.
    Предоставляет функционал для:
    - Управления профилями преподавателей
    - Управления курсами и их экземплярами
    - Фильтрации данных
    - Распределения преподавателей по курсам
    - Парсинга Excel файлов
  version: 1.0.0
  contact:
    name: FAH Team
    email: support@innopolis.university
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.fah.innopolis.university
    description: Production server
...
paths:
...
  /allocation/allocate:
    post:
      tags:
        - Allocation
      summary: Распределение преподавателя на курс
      description: Назначает преподавателя на экземпляр курса с указанной позицией
      operationId: allocateFaculty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AllocateFacultyRequest'
      responses:
        '200':
          description: Преподаватель успешно распределён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllocateFacultyResponse'
        '400':
          description: Некорректные входные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== PARSING ====================
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Сообщение об ошибке
      required:
        - error

    # ==================== PROFILE SCHEMAS ====================
...
    AllocateFacultyRequest:
      type: object
      properties:
        instance_id:
          type: integer
          format: int64
          description: ID экземпляра курса
        profile_id:
          type: integer
          format: int64
          description: ID профиля преподавателя
        position_type:
          type: string
          description: Тип позиции (PI, TI, TA)
          enum:
            - PI
            - TI
            - TA
        groups_assigned:
          type: integer
          format: int64
          nullable: true
          description: Количество назначенных групп
      required:
        - instance_id
        - profile_id
        - position_type
    AllocateFacultyResponse:
      type: object
      description: Пустой объект при успешном распределении

    # ==================== PARSING SCHEMAS ====================
...